Add base class for OIDCAuthenticationToken
