+++ iam-login-service/src/main/java/it/infn/mw/iam/authn/oidc/PendingOIDCAuthenticationToken.java
@@ -16,64 +16,27 @@
 package it.infn.mw.iam.authn.oidc;
 
 import java.util.List;
-import java.util.Map;
-
-import org.springframework.security.authentication.AbstractAuthenticationToken;
 
 import com.nimbusds.jwt.JWT;
 
 @SuppressWarnings("java:S2160")
-public class PendingOIDCAuthenticationToken extends AbstractAuthenticationToken {
+public class PendingOIDCAuthenticationToken extends BaseOIDCAuthenticationToken {
 
   private static final long serialVersionUID = -3039943580483543553L;
 
-  private final Map<String, String> principal;
-
-  private final String subject;
-  private final String issuer;
   private final transient OIDCProviderMetadata serverConfiguration;
-  private transient JWT idToken;
-  private final String accessTokenValue;
 
   public PendingOIDCAuthenticationToken(String subject, String issuer,
       OIDCProviderMetadata serverConfiguration, JWT idToken, String accessTokenValue) {
 
-    super(List.of());
-    this.principal = Map.of("sub", subject, "iss", issuer);
+    super(subject, issuer, List.of(), idToken, accessTokenValue);
 
-    this.subject = subject;
-    this.issuer = issuer;
     this.serverConfiguration = serverConfiguration;
-    this.idToken = idToken;
-    this.accessTokenValue = accessTokenValue;
 
     setAuthenticated(false);
-  }
-
-  @Override
-  public Object getCredentials() {
-    return accessTokenValue;
-  }
-
-  @Override
-  public Object getPrincipal() {
-    return principal;
-  }
-
-  public String getSub() {
-    return subject;
-  }
-
-  public JWT getIdToken() {
-    return idToken;
   }
 
   public OIDCProviderMetadata getWellKnownEndpoint() {
     return serverConfiguration;
   }
-
-  public String getIssuer() {
-    return issuer;
-  }
-
 }
